Scoped Slots
Slots can be scoped. This means that you can pass data from where the
slot was defined to where the markup that will be rendered in the slot
is defined.
We bind props to the slot definition and when we define the markup to be
rendered these props will be available there.
To have access to this data we can surround the markup with a template tag
and add the v-slot directive, passing the name of the slot, and attributing
a value to it, this value is  the object that will hold all the data from
the slot definition.