Keep alive
When we remove and element from the DOM all data is lost.
for example, we have an input field and add some value to it, if we render
another componente in the input`s place, and bring the input back, the
data inserted previously is lost.
We use the <keep-alive> tag to surround elements we want/need to save the
current state, even if the component is removed from the DOM. When we bring
the component back all information stored is kept.